requirements.txt
pyqt6==6.9.1
pyqt6-qt6==6.9.1
pyqt6-sip==13.10.2
pyserial==3.5
